requirements.txt
# Core Django
Django>=6.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Environment Variables (for local/production settings)
python-decouple>=3.8

# Image Processing (optional but recommended)
Pillow>=10.0.0

# Cloud Storage (for production - choose one)
# Option 1: AWS S3 (most popular)
# django-storages>=1.14.0
# boto3>=1.28.0

# Option 2: Cloudinary (easier for beginners)
# cloudinary>=1.36.0

# Database (for production - uncomment if using PostgreSQL)
# psycopg2-binary>=2.9.0

# Security (for production)
# django-cors-headers>=4.3.0  # If you need CORS support
